Minor fixes to logging and spelling.
Fixed spelling of "access" (was "acess") in "strings.xml".

Removed code that dereferenced a variable within an "if" statement that
already checked the variable was always null. The code caused a stack
trace for a "NullPointerException" to be logged every time. The value of
the "openExportImportDialog" field was always "false", so the code in
"onResume" that depended on it being "true" was also removed.