Added new "Session Best" statistics.
The "Session" statistics showed the average-of-N times for the most
recent N solves in the current session. This made it impossible to tell
if the "Best" statistics were achieved in the current session or not. A
new split into "Session Best" and "Session Current" (perhaps that could
be "Session Latest"?) shows both the averages of the most recent solves
and the best averages for any N consecutive solves in the current
session. Note, though, that the "All Time Best" includes the solves from
the current session, so it may show a new best average-of-N solves where
that sequence of solves is spans into the current session, so the
solution is not perfect. There is no space for more columns to get
around that fact.

A new (well-tested) "AverageCalculator" class is used to calculate the
averages, taking the nasty details away from the DB code. A "Statistics"
class aggregates multiple "AverageCalculator" instances to allow all
statistics to be calculated from a single database read of all solve
times, past and present. This is simpler and faster than the old
approach, which did multiple full DB reads and partial DB reads, one
read for each statistic. These new classes replace the code for
calculating the statistics for the graph/statistics tab. They could be
used in future for other statistics, such as those generated when
sharing statistics, or those shown on the main timer tab.

Redundant declaration of the statistics table layout in separate layout
files for portrait and landscape has been reduced to a single new layout
file included into both layouts. The layout has also been simplified,
but there was very little impact on the related code.

The solve count formatting was changed to add appropriate grouping
separators for the user's locale. This makes it a bit more readable when
the count goes above 1,000, as the value gets a comma (at least for
English locales).

The positioning of the "refresh" button on the statistics table is
currently wrong. It is aligning to the top of the table, not the center.
That will be fixed in another commit before this topic branch is
complete.