Added average-of-N lines to graphs.
Added average-of-N lines to the graphs on the graph tab. For current
session, the averages are Ao5 and Ao12. For all times, the averages are
Ao50 and Ao100. A legend is shown to explain the meaning of the
different lines.

Some changes to the status-quo have also been made: the line for all
times is made thinner if times from all sessions are shown, as this will
make it easier to see finer details when lots of times are bunched
together; the line is no longer dashed, as it looks cleaner and makes
sure that the peaks on the graphs are presented accurately (i.e., they
do not fall on a gap in the line between the dashes and fall short of
where they should be); and the graph of all times includes the times
from the current session, which is consistent with the way the
statistics table displays the all-time best statistics.

The chart data sets for the lines are all computed from a single
database read, the same as they are for the statistics table. It might
be possible to combine both reads into one, but that is not attempted at
this time.

The summary statistics on the "TimerFragment" have been ported to the
new "Statistics" class, so they are calculated on a single read. This is
not extended to the computing of the all-time best and worst times,
though.

Some tentative Google translation of "Everything" (for use in the
legend) into Portuguese and Czech needs to be reviewed by someone
competent.

The line colors for the averages are temporarily hard-coded. They will
need to be made theme-specific to ensure that they do not disappear
against the background color.

Changes are not yet complete: the "mean" line seems to be broken; the
positioning of the refresh button on the statistics table is still
wrong; and the colors need to be handled properly (see above).