Support for "DNF" in text import/export and improved import speed.
Added a trailing, optional "DNF" field to the supported format of simple
text files used for import/export. The inability to export "DNF" solves
made the text format useless when I tried exporting 1,000+ solves to
text (on a real device) and then importing them (on an emulator). All of
the statistics were a mess, as DNF solves were treated as normal solves
with ridiculously high and low times, which mess up the averages, too.

Perhaps the "DNF" field in the text file should replace the first time
field (and the time can then be recorded as 0.00 with the penalty flag
set to DNF), but it is added as a fourth field in this tentative
implementation. The importing operation is still compatible with the
older format. If an old version of the app attempts to import the new
text format with "DNF" solves, those solves will be ignored and reported
in the count of errors. In terms of the integrity of the statistics,
this is probably no better or worse than importing DNF solves as
successful solves, which was the old behaviour.

The "PuzzleUtils.parseTime" method was very slow. On importing 1,077
puzzle results from a text file (on an emulator), it took 8.3 seconds to
parse the data and then 1-2 minutes to insert it into the database. Of
the 8.3 seconds spent parsing, 8.0 seconds were down to "parseTime". I
measured the performance of the new version and it is 217 times faster!
Parsing 1,077 lines now takes about 0.3 s in total (before the database
insertions start). This avoids a long time looking at the progress bar
sitting at zero before progress starts being reported for the database
insertions.

The 1-2 minutes to perform the database insert has been reduced to about
2 seconds. The inserts are now wrapped in a single database transaction,
so there is no journal file overhead after each insertion. The bulk
delete operations now work the same way. Progress reporting is
integrated into the new database operations. It is not used for the
deletes, but it could be in the future if required.