Added line chart for best time progression.
Added a line to the chart on the graph fragment to show the progression
of best solve times. Each time a new best time is set, a new point is
added to the line and marked with a circle. These circles are then
linked together with by the line, which can only descend from left to
right, never rise.

As with other commits, the color of the line is not yet set cleverly
enough to avoid clashing with the background color from the theme.

As this change is relatively minor, I will merge it with the other
changes to the graphs and statisitcs on the "session_stats" branch and
manage all of those changes together.