Fixed crashes due to fragments closing the database.
The graph tab would usually crash on rotating the device because of some
sort of race condition with one fragment closing the database and
another trying to access it. The "SQLiteOpenHelper" (the super-class of
"DatabaseHandler") keeps a single, cached instance of the database that
is shared across all instances, so closing it causes all sorts of
problems and crashes.

There is really no need to close the database connection; it can be kept
open for the life of the application and Android/Linux will close it
properly when the application process exits. Therefore, a single
instance of the handler and its cached database instance are created in
"TwistyTimer" (the "Application" object) and these are accessed when
needed. Activities and fragments do not need to cache the instance of
the handler or the database and they never need to close it.

The code is simpler, the database will be closed automatically by the
system when the time comes, and the graph tab does not crash any more.
Result!