Improved statistics loading and graph of "best" times.
The changes to the way the statistics are loaded and calculated have
made the loading much, much faster. The "manual" button to refresh the
statistics seemed out of place and unnecessary. The statistics are
loaded in about the same time as the graph, and that has no refresh
button, so I removed the refresh button for the statistics. The progress
indicator is still retained, though.

The graph of the "best" is (expected to be) very sparse, as "best" times
are only set occasionally. Therefore, if the graph is showing 500 or
less times (zoom in if there are more than 500), then the values of the
"best" times identified among those 500 (or less) times will be shown on
the graph beside the circles marking those "best" times. If there are
more than 500 times visible, the "best" time values are not drawn.

Updates to the versions of the support libraries caused problems with
the tab icons. These have been fixed.