Line colors on charts are now taken from the theme.
Line colors on the charts are now selected to ensure that they always
contrast with the background color, regardless of the theme. Support for
up to three average-of-N lines is implemented, but more could be added,
just requiring extra theme attributes to define the colors. At present,
only two average lines are drawn, so adding one more is trivial.

Refactored the "values/styles.xml" and "values-v21/styles.xml" to reduce
redundant declarations and to make things easier to maintain. For the
most part, the attributes for the color theme styles are now only
defined once.

Updated some Gradle versions on command from Android Studio.