Moved position of best time values in chart to below data point.
The default position of the value text for the best times was above the
data points, which looked bad because the "all times" line always
crossed over the text. The text value is now shown below the data point
instead. There did not seem to be any way to configure this directly
through the chart API, so a new renderer was required.

Made all lines the same width in the session graph and all the same
width (but a bit thinner) in the full history graph.

Allowed the best time values to be drawn when there were more total data
points in the visible chart area.

Removed some stray logging code that found its way into the previous
commit.