Chart best AoN values and localise date format for X-axis labels.
The single best average-of-N value for each "N" is now marked on the
chart using a circle over the data point. Only a single value is shown;
they are not linked together in a progression the way the best
(non-average) times are, as that would probably make the chart too busy
looking. The value of the best AoN time is also shown on the graph. As
the best AoN requires its own data set, the legend needed to be
customised to avoid duplicated entries: one for the AoN line and one for
its corresponding best-AoN "line" (just one point, really).

Added some localisation of the dates along the X-axis, so that, in the
US, the format is "MM/DD" instead of "DD/MM" (which most of the rest of
the world use). The formatting of these values as the chart is being
populated with entries is now also more efficient.