Fixed issue with incorrect legend rendering.
The legend was being drawn incorrectly due to some cached measurements
being held by the Chart library. The order of the setting of the chart
data and the custom legend values was the problem. To avoid future
problems, these call have been moved from the fragment into the
"ChartStatistics" class to ensure they are set correctly. The limit line
for the mean value was also moved in, as it makes more sense to keep all
of the line creation and configuration in the same place. A few other
related settings were moved in, too, to de-clutter the fragment class.