Improved management of chart and statistics background tasks.
The background tasks for the chart and statistics are now deferred
unless the tab is visilbe. This prevents them from slowing down the
start of the application which typically shows the time tab first and
generates a scramble in the background. Once the user switches to the
graph tab, any pending tasks, such as those caused by adding a new time
or toggling the "history" switch, will be executed immediately. The new
approach to calculating the graph and other statistics is faster, so
there is very little reason to prepare the results in advance of viewing
the respective tab.

Drawing the values on the chart for the best AoN time for each N did not
look good, as there was too much clutter around them. Only the circles
are now shown. The values can be read from the statistics table, as they
are already present there.

Removed some more stray logging code from the previous commit.