Clean-up of broadcast/receive code. Upgrade of Butterknife.
The hard-coded strings made the broadcast/receive code brittle, as a
spelling mistake could introduce a bug. Using constants makes it much
easier to manage and makes it easy to locate references to the constants
to see which receivers handle which broadcasts.

Using the Intent action in conjunction with an Intent extra named
"action" was a bit confusing. The new approach uses Intent categories to
indicate if an Intent relates to a UI interaction or a change to some
data, etc. The normal Intent action is then used to specify the action.

Some code used "Broadcaster" and some code did not. The new constants
for categories and actions, and the broadcast utility methods are now in
a new TTIntent class. A wrapper class for use in Fragments avoids the
need to protect the handling code in a test of "Fragment.isAdded()",
makes it clear what category of actions will be received at the point
that the receiver is created (the category is a parameter to the
constructor), and automatically creates the IntentFilter when a receiver
is registered.

An upgrade to Butterknife 8.x makes the code that binds the "sticker"
views much cleaner, as views can be bound to an array and these can be
operated on with a Butterknife.Action. One or two misplaced or missing
"unbind" operations were fixed in the Fragments.

A static utility method to get the application context from the
"TwistyTimer" class makes it easier to get a context that is safe to use
without worrying about memory leaks. As it is an application content, it
is not useful for getting theme attributes, but it is fine for most
other things.

A few hacks to provide the "this" of the outer class to an inner class
were changed. Fields are unnecessar; just use "OuterClass.this" from
the inner class.

The new statistics collectors, etc. have been moved into a "stats"
package. Loaders will be added there in a future commit. Cleaning up the
broadcast/receive code was motivated by the need to ensure that it can
be integrated into new Loaders for the statistics and chart without
making it even more brittle.