Fixed issues with crashes and fragments starting twice.
When the app was started, problems with the initialisation of the puzzle
type spinner caused it to fire "onItemSelected" during layout. At this
point, the various fragments were already created and were running, or
had already run, various Loaders and AsyncTasks. However, the stray
event caused all fragments to be destroyed and re-created. This was
happening on every start-up of the application and every time the device
was rotated, etc., which made start-up much slower than it needed to be.
A small change to the set-up of the spinner (documented in the code,
lest it be forgotten) fixed this problem....

.... However, this uncovered a hornets' nest of other problems! Because
the fragments were not being re-created, they were preserving their
state across configuration changes. However, this state was not being
restored sufficiently. In particular, the references held in the
"current*FragmentInstance" fields of "TimerFragmentMain" were being lost
and this was causing crashes (NPEs).

Restoring these references was possible by overriding
"NavigationAdapter.restoreState", but that was not enough. More
problems were being caused, mostly in "TimerFragmentMain.setupPage",
because it was calling methods on the fragments without knowing what
state each fragment was in. In particular, the fragments very often had
not completed "onCreateView" when they were being asked to hide or show
views. The main fix for this was to override "setUserVisibleHint" in the
subordinate fragments to allow them to handle the transition to or from
the fragment themselves. For example, when the "TimerListFragment" comes
into view, it will show the FAB and when it is replaced by a different
fragment (on switching tabs) it will be notified again and will hide the
FAB. This allows it to check all the necessary state and make these
changes safely and in the "setUserVisibleHint" code in its own class.

Another complication was the "onBackPressed" event handling. Again, this
required "TimerFragmentMain" to drill-down into details of the other
fragments and this frequently caused crashes after an orientation
change. Local broadcasts and receivers were used to handle this event,
but there were not suitable for a very simple reason: they were handled
independently. When one fragment needed to "consume" the back button,
for example, to close the FAB sheet on the "TimerListFragment", it was
not simple to notify other receivers of the broadcast that the event had
been consumed. Chaining broadcasts together was attempted in the old
code, but this was a confusing workaround for the underlying problem.
Instead, then, the new approach does not use broadcasts to propagate the
"onBackPressed" event; it passes the event down to the lowest level and
then only proceeds to propagate the event at a higher level if it has
not already been consumed. Simply, the event is passed down to each
fragment via "onBackPressedInFragment" and as soon as the first such
call returns "true" to indicate the event is consumed, propagation of
the event stops. The event is passed down initially from "MainActivity",
so if the event gets back there without being consumed, the app exits.
The result is the same behaviour as before, but it is easier to follow
and each fragment can handle the event in a safe manner, e.g., ensuring
that views are created before attempting to modify them, etc. This fixed
a bunch of other crashes.

"TimerFragmentMain" no longer needs to hold references to the other
fragments, so this makes save and restore of state simpler (no need to
keep the "NavigationAdapter.restoreState" method that I had used as a
stop-gap).

When the device is rotated, it now stays on the tab that it was showing.
This fixes the problem I had when viewing the graph and turning the
phone to see it in landscape: the timer tab was being selected and I had
to switch back to the graph tab. Again, this required a good few minor
tweaks to the use of the "currentPage" field, etc.

The code is now "tricked out" with debug logging in major life-cycle
methods. Each log statement is guarded by a "DEBUG_ME" flag which can be
set to "true" or "false" at the top of each class. As the flag is
static, the guarded code will probably be stripped out by ProGuard when
it is set to "false", so there is no overhead in leaving the code in
place and it makes debugging life-cycle problems much easier. The log
"TAG" fields use "SomeClass.class.getSimpleName()" to ensure that if the
name of a class changes, the log tag will be changed, too.

As I am using the current stable release of Android Studio, I upgraded
the Gradle version and Android Gradle Plug-in version to the ones
Android Studio was insisting upon.