Ported chart and statistics tasks to Loader API.
The chart and statistics AsyncTasks are now AsyncTaskLoaders that are
managed by the LoaderManager for the TimerGraphFragment. Monitoring of
broadcast intents is integrated into the loaders. This makes things much
easier to manage, particularly when coordinating things with the
life-cycle of the fragment.

The loaders are started in "onActivityCreated", as this guarantees that
the Fragment.onCreateView method has completed and that any results
being delivered will have Views ready to display them. However, there is
still a test of "Fragment.isResumed" from "onLoadFinished", just for
safety, as a load might finish after "onDestroyView".

Because there are two loaders and they return different types of data,
separate "LoaderCallbacks" implementations (in anonymous inner classes)
were used. This actually makes things a bit easier to comprehend, as
there is no need for shared methods that "switch" by the loader ID.

A significant optimisation has been made in the loaders: if a single new
time is added to the session, it is added to the existing Statistics and
ChartStatistics objects held in the loaders without requiring any new
access to the database. This "fake" load is delivered to the fragment
like any other, but a "Wrapper" is required around the existing
Statistics or ChartStatistics object to allow the wrapper to be a new
instance and trick the LoaderManager into delivering the same "payload".
(Ordinarily, the LoaderManager will not call "onLoadFinished" if the
data it is delivering is the same object as in the previous delivery.)

Lots of issues with the new TTIntent stuff (previous commit) have been
sorted out.

ButterKnife was upgraded from 7.x to 8.x to fix some issues. This also
made it possible to handle the bindings for the cube "stickers" in a
much neater way, as ButterKnife 8 allows binding of an array of view IDs
to an array of View objects.

There are still lots of start-up problems in TimerFragmentMain that are
causing crashes, but these have been (mostly) fixed on the
"fix_double_start" branch, so that branch will be merged into this
"session_stats" branch before going any further.

An optimisation from a previous commit that deferred loading of the
chart and statistics data until the tab became visible has been removed.
It was a bit messy to integrate with a Loader and, with improvements to
the start-up sequence fixed on "fix_double_start" and (see below) the
fact that the statistics will be shared across fragments, it was not
worth the effort in preserving this. However, it may be possible to
re-introduce it later by kicking off the chart statistics loader from
the "TimerGraphFragment.setUserVisibleHint" method (if it has not
already bee started). Time will tell.

The next step will likely be to move the "StatisticsLoader" under the
control of "TimerFragmentMain" and then, from there, distribute results
to "TimerFragment" for its summary statistics on the timer tab and to
"TimerGraphFragment" for its full statistics table.