Share statistics between fragments.
A single loader for the statistics is managed by TimerFragmentMain and
these statistics are pushed to a cache. Other fragments can then access
the cache and register themselves as observers for later updates. This
allows the timer and graph fragments to share a single "Statistics"
instance, so only one load from the database is required to service both
fragments.

TimerFragment now uses the data from the shared Statistics to manage the
notification of best or worst record times. These records are declared
if the user has recorded four previous non-DNF times. This is different
from before, where the user had to record four previous times in the
current session, which did not seem to be right.

TimerListFragment grabs the statistics, too. It will use these if the
user shares averages, but that will be in the next commit.

The unexplained fiddling with the layout of the chart in landscape mode
has been restored. I removed in an recent commit because it caused NPEs,
but added it back, with more null-checks, because I discovered why it
was there! However, I removed the scrolling, which looks really bad if
the device is rotated to landscape from portrait while viewing the
chart.