Statistics used by TimerListFragment for sharing averages.
TimerListFragment now uses the "Statistics", loaded once in
TimerFragmentMain and distributed through StatisticsCache, to generate
the average-of-N and histogram strings for sharing scores. The code in
PuzzleUtils has been updated to match. "Statistics" and
"AverageCalculator" do all the necessary heavy lifting to produce the
information required. No database reads are required, other than that to
update the Statistics, which would be happening anyway.

The changes respect the previous behaviour. However, it would not be
difficult to expand things to share a histogram of all times (full
history), or to add some option to graph the histogram within the app.

Lots of small clean-up were done after running some code inspections.
This gets rid of lots of warnings, but does not change any behaviour.