Fixes to transitions and layouts. Got carried away.
Fixed problems with the transitions into and out of the
SettingsActivity. Merging the two settings activities made this much
easier. There was no way I could see to use "PreferenceScreen" in the
XML to nest the timer appearance settings and then use one fragment, as
it always caused a dialog fragment to open that obscured the navigation
tool-bar. Still, things are more compact, now. The main problem was that
the first settings fragment put itself on the back-stack, so a double
back press was required to exit settings. Then, there was a big issue
with the life-cycle of the MainActivity when it was recreated. Cleaned
up the recreation of the MainActivity, posting "Activity.recreate()"
works very nicely, now.

Changed much of the use of hard-coded strings to specify preference keys
to use string resources instead. A new "Prefs" class makes this very
easy to use. It adds some safety, as misspellings are not a problem, and
it makes it easier to find things in the code. The resource ID names do
not have to be the same as the key names, so they can be more
meaningful.

The Export/Import dialog stuff was making MainActivity very confusing,
so I re-jigged it to use MainActivity only to relay information between
the fragments, not to try to store any of it. I makes things easier to
follow, as it is all controlled by ExportImportDialog and simply
executed from MainActivity.